Moved the `get_all_pod_templates/get_generable_templates` logic outside the `DocumentGeneratorLinksViewlet` so it is easier to override. It is now an `IGenerablePODTemplates` adapter.
